Removed obsolete content and restructured
