--- iBEAt/iBEAt_models/README.md
@@ -1,5 +1,0 @@
-# weasel_apps/iBEAt_models
-
-A collection of weasel applications for iBEAt.
-
-If you wish to update the "iBEAt_Model_Library", simply run `git submodule foreach git pull update --remote --merge origin master` in your terminal.
