--- iBEAt/iBEAt_models/iBEAt_Model_Library/.vscode/settings.json
@@ -1,7 +1,0 @@
-{
-    "git.ignoreLimitWarning": true,
-    "python.linting.pylintEnabled": true,
-    "python.linting.enabled": true,
-    "python.linting.pylintPath": " C:\\Users\\medkshaa\\appdata\\roaming\\python\\python36\\site-packages",
-}
-
