--- iBEAt/iBEAt_models/iBEAt_Model_Library/README.md
@@ -1,4 +1,0 @@
-# iBEAt_Model_Library
-
-Core image processing functions used in iBEAt pipeline
-Python Library of model functions
